NeoVim Statusline Update + Select Mode
Statusline now updates on 'ModeChanged'. This makes the statusline
actually show which mode you are in.

Made it so that select mode can be entered with the mouse (not oft used)
and Shift + unprintable movement (also not oft used). Should add a bit
of extra functionality instead of things I don't need/use.